requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
twilio>=8.0
Pillow>=10.0
PyJWT>=2.8
channels>=4.0
channels_redis>=4.1
boto3>=1.28
django-storages>=1.14
openai>=1.0
python-dotenv>=1.0.0
daphne>=4.0.0
psycopg2-binary>=2.9.0
django-simple-history>=3.4.0
